RoomMesh überarbeitet
* Türen werden nun aus dem Mesh ausgestanzt
* Es gibt einen Bug, der das Mesh zum flackern bringt und eine
horizontal zu große lücke ausstanzt (problem beim zusammenfassen?)
*Die Ausrichtung der Triangles ist fehlerhaft
*Boden und Decke fehlen

+++ Assets/Scripts/AbstractObjects/StickTo.cs
@@ -76,7 +76,7 @@ public class StickTo : TransformingProperty {
 		}
 	}
 		
-	public override int ExecutionOrder{
+	public override float ExecutionOrder{
 		get { return 5; }
 	}
 }
